api/requirements.txt
boto3==1.9.237
django-ckeditor==5.7.1
django-filter==2.2.0
django-health-check==3.11.0
django-storages==1.7.2
Django==2.2.5
djangorestframework==3.10.3
Faker==2.0.2
gunicorn==19.9.0
markdown==3.1.1
pillow==6.2.0
psutil==5.6.3
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0